moved common functions to a new file
